src/requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==2.2.16
django-debug-toolbar==3.1.1
django-mathfilters==1.0.0
django-sass==1.0.0
django-tinymce==3.1.0
django-widget-tweaks==1.4.8
flake8==3.8.4
idna==2.10
libsass==0.20.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
pip-tools==5.3.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
pytz==2020.1
regex==2020.10.23
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
whitenoise==5.2.0
